+++ front_end/app/components/Editor/Modal/searchResultModal.tsx
@@ -1,4 +1,5 @@
 import { useState, useEffect } from "react";
+import useHoverModal from "../Hooks/useHoverModal";
 import ModalOverlay from "./modalOverlay";
 
 interface SearchResultModalProps {
@@ -26,9 +27,7 @@ const SearchResultModal: React.FC<SearchResultModalProps> = ({
 }) => {
   const [render, setRender] = useState(false);
   const [animate, setAnimate] = useState(false);
-  const [selectionModalOpen, setSelectionModalOpen] = useState(false);
-  const [selectionModalPosition, setSelectionModalPosition] = useState({ x: 0, y: 0 });
-  const [selectedWord, setSelectedWord] = useState("");
+  const { selectionModalOpen, setSelectionModalOpen, selectionModalPosition, selectedWord, handleWordHover } = useHoverModal();
 
   useEffect(() => {
     if (isOpen) {
@@ -40,23 +39,6 @@ const SearchResultModal: React.FC<SearchResultModalProps> = ({
     }
   }, [isOpen]);
 
-  const handleWordHover = (event: React.MouseEvent<HTMLElement>, word: string) => {
-    const subModal = document.getElementById("subModal");
-    const subModalWidth = subModal?.offsetWidth || 200;
-
-    const hoverRect = (event.currentTarget as HTMLElement).getBoundingClientRect();
-    const offsetX = 100;
-    // サブモーダルをリストアイテムの左側に配置するためのX座標を計算
-    const modalPositionX = hoverRect.left - subModalWidth - offsetX;
-
-    setSelectionModalPosition({
-      x: modalPositionX + window.scrollX,
-      y: hoverRect.bottom + window.scrollY,
-    });
-    setSelectedWord(word);
-    setSelectionModalOpen(true);
-  };
-
   const handleClose = () => {
     setSelectionModalOpen(false);
     onClose();
@@ -67,23 +49,23 @@ const SearchResultModal: React.FC<SearchResultModalProps> = ({
   const { width, left } = editorPosition;
 
   const modalContentStyle: React.CSSProperties = {
-    position: 'fixed', // Updated to fixed to position relative to viewport
+    position: 'fixed',
     top:  position ? `${position.y}px` : '50%',
     left: `${left}px`,
     width: `${width}px`,
-    transform: position ? 'none' : 'translate(-50%, -50%)', // Adjusted for custom position
+    transform: position ? 'none' : 'translate(-50%, -50%)',
     opacity: animate ? 1 : 0,
     transition: 'opacity 0.3s ease-out',
     backgroundColor: '#fafaf9',
     padding: '20px',
     borderRadius: '4px',
     boxShadow: '0 2px 10px rgba(0, 0, 0, 0.2)',
-    maxHeight: '50vh',
+    maxHeight: '40vh',
     overflowY: 'auto',
     zIndex: 1050,
   };
 
-  const selectionModalStyle = {
+  const selectionModalStyle: React.CSSProperties = {
     position: 'fixed',
     top: `${selectionModalPosition.y}px`,
     left: `${selectionModalPosition.x}px`,
@@ -91,7 +73,7 @@ const SearchResultModal: React.FC<SearchResultModalProps> = ({
     boxShadow: '0 2px 10px rgba(0,0,0,0.2)',
     padding: '10px',
     borderRadius: '4px',
-    zIndex: 1060, // Ensure it's above the rhyme search modal
+    zIndex: 1060,
   };
 
   return (
@@ -106,6 +88,7 @@ const SearchResultModal: React.FC<SearchResultModalProps> = ({
           {searchResults.map((result, index) => (
             <li
               key={index}
+              onMouseOver={(e) => handleWordHover(e, result)}
               onClick={() => onWordSelect(result)}
               className="hover:bg-indigo-100 cursor-pointer"
             >
@@ -114,6 +97,11 @@ const SearchResultModal: React.FC<SearchResultModalProps> = ({
           ))}
         </ul>
       </div>
+      {selectionModalOpen && (
+    <div style={selectionModalStyle}>
+      <p>{selectedWord}</p>
+    </div>
+  )}
       </>
     );
 };
