+++ front_end/app/components/Editor/editor/page.tsx
@@ -138,7 +138,6 @@ const MyEditor: React.FC<MyEditorProps> = () => {
     fetchNotes();
   }, []);
 
-
   const titleActions = titleKeyActions({
     textareaRef,
     onEnter: () => {
