+++ front_end/app/globals.css
@@ -12,7 +12,3 @@
   .bpsgbes {
     @apply bg-white  border  shadow-lg;
   }
-
-  .no-transition {
-    transition: none !important;
-  }
